auto_creative_engine/requirements.txt
torch
diffusers
transformers
accelerate
rembg
Pillow
openai
groq
streamlit
huggingface_hub
